src/requirements.txt
assemblyai
python-dotenv
tk
openai